xrootd4j: fix how we turn off signing policy if TLS is on
Motivation:

Signed hash verification is unnecessary if TLS is on,
whether before or after authentication.

Hence, we should not tell the client to turn it on
if going to TLS is being or will be recommended.

For the TPC client, life is simpler.  We simply
rely on the source to tell us whether to verify
signed hashes or not.

Modification:

To be consistent, we must do this immediately
in the protocol response.  If the server "uses"
TLS ... which indicates either an instruction
to go to tls before login or after, the sigver
settings are overridden with a default object
which turns them off.

Result:

The server will not erroneously require the
client to send the signed hashes.

Our TPC client simply obeys the source
end directive.

Target: master
Request: 4.0
Request: 3.5
Patch: https://rb.dcache.org/r/12915/
Acked-by: Tigran