xrootd4j: add RequiresTLS interface and method for checking if TLS is already on
Motivation:

The ZTN protocol specifies it should fail if TLS is not already
negotiated between client and server.  We need a way of enforcing
this early check.

Modification:

Add marker interface indicating the handler requires TLS.
Add method for checking for the SSLHandler in the pipeline.
Use these together to fail if the handler requires TLS
and the SSLHandler is absent.

Result:

ZTN requirement is supported.

Target: master
Patch: https://rb.dcache.org/r/12909/
Acked-by: Tigran