xrootd4j: move TLS (mostly) from dCache to xrootd4j
Motivation:

Except for a very small implementation-specific method
providing the SLL context, the TLS-Netty functions can
all be moved into the library for general availability.

Modification:

Create an abstract handler, allow the the factory
to instantiate it on the basis of a property
giving the class FQN.

Result:

Only a residual class is left for dCache to implement.

Target: master
Patch: https://rb.dcache.org/r/12916/
Acked-by: Tigran