xrootd4j: add scitokens authz plugin module
Motivation:

With the exception of how the validation step is done
(which depends on the implementation: in dCache, it is
done through login/gPlazma), the entire scitoken/JWT
authz stack actually should be in the xrootd4j library.

Modfication:

Most of the code that was in dCache has been brought
into the new module.

A token validator interface has been created, and
the factory made abstract.  Thus the user of
the library needs to extend the abstract factory
to provide the validator, and implement the
latter.

Result:  Scitoken authorization is mostly defined
in the library.

Target: master
Patch: https://rb.dcache.org/r/12913
Acked-by: Tigran