xrootd4j: use CGI tpc.dlgon to determine if delegation is on
Motivation:

For xrootd TPC there are two protocols, one which uses the
rendez-vous key, and one which relies on delegated credentials.
In the latter case, various tpc CGI elements should not be
sent by the TPC client to the source, or it will get
confused and think the initiating client needs to call
open as well (which it does not when delegating).

Currently, our TPC client filters out those elements
by checking some ad hoc properties the GSI module adds
to the authentication context.

This solution, however, is not standard and would have
to be reimplemented for each new protocol in which
delegation is possible (like Scitokens).

Fortunately, since the original TPC protocol was designed,
a new CGI, "tpc.dlgon", was added with value either 0 (OFF)
or 1 (ON), which allows the destination server to
tell the TPC client what to do.

Modification:

Eliminate the ad hoc solution and add a check of tpc.dlgon.

Result:

Correct behavior not only for GSI but automatically for
any protocols called with "--tpc delegate only" on the
client command line.

Target: master
Request: 4.0
Request: 3.5
Patch: https://rb.dcache.org/r/12935/
Acked-by: Tigran