xrootd4j: move GSI-specific security info access out of generic client handler*
Motivation:

Not all sec strings are constructed as key-value pairs.
For instance, the ztn string has this structure:

&P=ztn,flags:maxtsz:[tokloc[tokloc[...]]]

Furthermore, the keys "v", "c" and "ca" are specific
to the GSI protocol and do not belong in the abstract
handler.

Modfication:

Instead of storing a map in the TPC client,
store the selected protocol info.

Remove reference to the GSI keys in the abstract
handler.  Refit GSI to use the SecurityInfo
object directly.

Result:

Code is correctly encapsulated.

Target: master
Patch: https://rb.dcache.org/r/12953
Acked-by: Tigran