xrootd4j: remove io.netty LogHandler from TPC client pipeline
Motivation:

It seems that io.netty logging does not play well
at the TRACE level with the turbo filter check of
trace enabled.  This means that the logger gets
added to the pipeline regardless.  The is very
costly in terms of bandwidth as the io.netty
logger does a hex dump of every message received.

Modification:

The TPC client really does not need to enable
io.netty logging under normal conditions,
so we can safely eliminate adding it to
the pipeline.

Result:

Bandwidth for TPC is restored to what is
reasonably expected.

Target: master
Request: 4.1
Request: 4.0
Patch: https://rb.dcache.org/r/13038
Acked-by: Dmitry
Acked-by: Lea