xrootd4j: switch to netty implementation to take advantage of openssl
Motivation:

The current code uses an older API for the SSL context
which does not allow for a pass-through to OpenSSL if
it is present.  The latter is the recommended way to
set up SSL/TLS in Netty, and is about 20% more efficient.

Modification:

Make the minor changes to accommodate use of the
Netty SslContext object insteal of javax.net.SSLContext.

Result:

The library now uses the Netty context which
checks for and uses the OpenSSL implementation
of TLS.

Note:  This patch breaks current dCache master.
Update should be applied only after the
corresponding changes are made to dCache.

Target: master
Patch: https://rb.dcache.org/r/13040/
Requires-notes: yes  (in dCache patch)
Requires-book: no
Acked-by: Tigran