xrootd4j-unix: revert credential to username
Motivation:

changed the credential sent from 'username' to 'unixUID GID'.
This was based on a slight misunderstanding of the vanilla
server code, on two counts (unix was an actual prefix of
the credential, not prepended to the buffer from the
credential type field; and UID GID in the code actually
stood for user and group ids, not names).

This caused problems for NDGF/Alice upon upgrade to this
version:

```
[ERROR] Server responded with an error: [3010] unix25021 25020 security protocol is not supported.
```

Modification:

Revert the code to what it was doing previously (sending
only the username).

Result:

UNIX protocol interaction (with EOS/xrootd server) should
work again.

Target: master
Request: 4.2
Request: 4.1
Request: 4.0
Patch: https://rb.dcache.org/r/13201/
Requires-notes: yes
Requires-book: no
Acked-by: Lea
Acked-by: Tigran