xrootd4j: fix two small bugs in third-party client authentication handling
Motivation:

The tpc client (embedded) has two small flaws which
were discovered once it was tested against the
change to support multiple authentication protocols
in the door.

Modification:

The Unix handler is sending the request with protocol
as empty string.  Evidently the vanilla server ignores
this, but we need it to be 'unix'.  And, that is what
the xrdcp client sends anyway.

Second, if one of the handlers that are added to
the pipeline fails, it must be removed or it
will try to reprocess the next authentication
reply as if it were intended for it (the response
does not include the protocol name, so the
only way to prevent this is to remove it).

The server handles this situation a little differently,
in that it only adds one handler at a time to the
pipeline, according to the client's request.

Result:

TPC from dCache to dCache works again.

Target: master
Patch: https://rb.dcache.org/r/13513/
Requires-notes: no
Requires-book: no
Acked-by:  Lea