xrootd4j: Enable support for authentication protocol chaining
Motivation:

The legacy dCache xrootd code requires only one authentication
plugin (aside from anonymous/no authn) to be loaded in the door.
This is not optimal, as it requires different endpoints to
support different authentication protocols in xrootd.

At the same time, the xrootd protocol specifies that the
server can give the client an ordered list of protocols
it supports and the client can try each of them.  dCache
should be able to support this.

Modification:

This patch sets up the framework for dCache to load
multiple plugins and support any of them on a
given connection.

The changes are as follows:

The XrootdAuthenticationHandler has been rewritten (modernized)
to subclass the XrootdRequestHandler. It is now responsible
only for authentication. It composes the following class:
The XrootdSessionHandler is responsible for managing login.
It subsumes some of the functionality of the previous
XrootdAuthenticationHandler, but adds the ability to try
one available authentication protocol after the next,
according the client's request, by adding and removing
the handler implementations from the Netty pipeline.
A few small changes have been made to support mapping of
protocols to their handlers.
The XrootdAuthorizationHandler now delegates the TLS check
to the specific handler class; this allows us not to
abort a non-TLS connection when tokens are not required
(strict=false).
In order to support anonymous authentication within the
authentication protocol chain, we need to map "none"
alternately to "unix", otherwise the client will give
up after the other protocols have failed. The uid:gid
credential it sends is just ignored by the NoAuthorization
handler.

Result:

A single door can now be configured to support all protocols
as an ordered chain.  An example config will be given in
the downstream patch that modifies the door code based on
the changes here.

Target: master
Patch: https://rb.dcache.org/r/13512
Requires-notes: yes (with the dcache-xroot patch)
Requires-book: yes (with the dcache-xroot patch)
Acked-by: Tigran