xrootd4j: put the attn response future on the client
Motivation:

To handle kXR_wait or kXR_waitresp, a scheduled future with
timeout is used.   This future is currently stored on the
handler that processes the request, which is incorrect,
as some handler in the pipeline preceding it could
process the follow-up response.

Modification:

Put the future on the tpc client, which is shared between
handlers in the channel pipeline.

Result:

Cancellation of the future works properly.

Target: master
Request: 4.2
Request: 4.1
Request: 4.0
Patch: https://rb.dcache.org/r/13517/
Requires-book: no
Requires-notes: yes
Acked-by: Dmitry