xrootd4j: set tpc info state to READY if updated when PENDING
Motivation:

The addInfoFromOpaque method supports the configuration
of the tpc info object by the merging of information
carried by the two clients in a TPC rendezvous call.
It is thus understood to be called twice, with its
state going from PENDING->READY, in whatever order
the clients update (unless cancelled or there is
an ERROR).

However, the state is currently not being
updated from PENDING when this happens.

Modification:

If the state is PENDING and update is called,
change this to READY (the update contract here
is understood to be that the info object is
updated twice in the case of a rendezvous tpc).

Result:

Correct semantics.

Target: master
Request: 4.2
Request: 4.1
Request: 4.0
Acked-by: Lea