xrootd4j: break write up into max frame size (serial) chunks
Motivation:

The native xrootd client uses an env variable, XRD_CPCHUNKSIZE,
to determine the size of the frame transmitted by the client
over the wire.  The client waits for the server to
consume all the bytes before considering the request successfully
sent.  This, however, does not determine how the server may
process this data internally.  Placing it all into a buffer
at once is currently the way dCache (through Netty) handles
this.

However, with writes we may encounter greater memory pressure
by doing this; in addition, it is quite possible for a client
to take down a pool on a write by simply (maliciously, perhaps)
increasing the chunk size to an enormous size greater than
what may normally be allocated to (direct) memory in the JVM
on the pools.

For this reason some (defensive) action seems to be in order.

Modification:

The solution adopted in this patch is to segment each write
request into chunks less than or equal to a maximum buffer
size, and to write this to disk serially, with a release
of that data buffer following each write.

To this end, the decoder and encoder have been modified
to handle incomplete or partial writes; the WriteRequest
constructor has been altered to accept parameters
corresponding to its packet parts (it is no longer
responsible for knowing how to read that data from
the buffer).

Result:

Where memory consumption before was hostage
to the client frame size, now the pool can be
configured for max buffer size; this can be
tuned to match the expected concurrency
(simultaneous connections).

Target:  master
Patch: https://rb.dcache.org/r/13639/
Requires-notes: yes (on the dcache patch)
Closes: #137
Acked-by: Tigran
Requires-book: yes (provided by #13637)