xrootd4j: disconnect if maximum frame size is exceeded
Motivation:

see #13639

This is the minimal version for backport to the
stable branches.

It uses a hardcoded maximum frame size and
disconnects if that is exceeded.

The client behavior will be to retry and
then finally give up, issuing a "file not found"
error.

Target:  4.4
Request: 4.3
Request: 4.2
Request: 4.1
Request: 4.0
Requires-notes: yes
Patch: https://rb.dcache.org/r/13647/
Acked-by: Dmitry