xrootd4j: correct regression in NoAuthenticationProvider
Motivation:

In order to support multi-protocol doors, the
unix protocol must be the last alternative sent
back to the client.

The strategy in dCache was to change the name
of the NoAuth handler to unix, but to continue
to allow gplazma:none to map to it.

Outside of dCache, however, the provider
loaded via SPI needs to know about both
alternative names in order to work, e.g.,
in the standalone server.

Modification:

Have the provider respond to either unix
or none, and change its name to unix.

Result:

The plugin can be loaded successfully
again outside of dCache.

Target: master
Request: 4.5
Request: 4.4
Request: 4.3
Request: 4.2
Patch: https://rb.dcache.org/r/13725/
Acked-by: Tigran