Motivation:
While of limited capability, the xrootd standalone server
can be useful for preliminary testing.

The server currently does not support the full range
of authentication/authorization capabilities; nor does
it support TPC.   One of the more important features
that might be useful to have, however, is TLS.

In order to integrate TLS, a few preliminary changes
are necessary.

Modification:

The following necessary changes are made here:
instead of injecting the vanilla XrootdAuthorizationHandler
into the DataServer channel pipeline, we use the XrootdSessionHandler
the way we do in dCache. For the moment, SigningPolicy and TLSSessionInfo
are unconfigured placeholders (TLS defaults to OFF and sigver to 0).
make the DataServerHandler subclass XrootProtocolRequestHandler
and remove the overridden doOnProtocol method in favor of the parent
implementation.
Make the default plugin in the option parser "authn:unix", so that
the server registers unix as its authentication protocol.

Result:

First step in integrating TLS into the standalone server.

Target: master
Requires-book: no
Requires-notes: no
Patch: https://rb.dcache.org/r/13728
Depends-on: #13725
Acked-by: Tigran